+++ NEWS.md
@@ -1,3 +1,11 @@
+v0.6.0 (2018-??-??)
+===================
+
+- WCSLIB has been upgraded to v6.2. Note that now the `obsgeo` field of
+  `WCSTransform` is a 6-element vector, instead of the previous 3-element
+  vector.  For backward compatibility, a 3-element vector can still be provided
+  and it will be automatically adjusted.
+
 v0.5.0 (2018-11-06)
 ===================